requirements.txt
apache-airflow[apache.spark]==2.7.1
apache-airflow-providers-openlineage>=1.8.0
apache-airflow-providers-apache-spark>=4.0.0
apache-airflow-providers-papermill
papermill
pyspark==3.5.1
pytest
requests==2.31.0
psutil
